// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  checkouts: typeof checkouts;
  credits: typeof credits;
  emailTemplates: typeof emailTemplates;
  feedback: typeof feedback;
  http: typeof http;
  polar: typeof polar;
  rateLimit: typeof rateLimit;
  resend: typeof resend;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;